# BackOffice/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API Projet-Cloud-S5",
        "description": "Documentation de tous les APIs de notre projet",
        "version": "1.0.0"
    },
    "paths": {
        "/api/inscription": {
            "post": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Inscription de l'utilisateur",
                "operationId": "d3d2dd59f195539f66a5649d47c47727",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "pseudo": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code PIN envoyé à user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "example": null,
                                            "nullable": true
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/validate-pin-inscription": {
            "post": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Valider le code PIN d'inscription",
                "operationId": "920a7ed84ff9a7a68845bc3925e5da0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "codePin"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "codePin": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code PIN envoyé à user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "example": null,
                                            "nullable": true
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "c7c2a052ce6bfef935122b55d9ac92ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "succes"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Code PIN envoyé à user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "example": null,
                                            "nullable": true
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Utilisateur",
            "description": "Opérations liées au contrôleur pour l'utilisateur"
        }
    ]
}